sort_rewriter: Don't attempt to contact the GPU if one isn't present.
Even when the compiler is built with CUDA, the compilation may not occur on a
machine with a GPU. Currently SortRewriter will always attempt to contact the
driver, which will fail in such scenarios. Let's avoid that at the admitted
runtime cost, especially since the error message ("invalid device ordinal")
is not particularly clear.

PiperOrigin-RevId: 736683228